Refactoring of publish script for clarity
