--- .eslintrc.json
@@ -1,20 +1,0 @@
-{
-	"globals": {
-		"artifacts": true,
-		"assert": true,
-		"contract": true,
-		"web3": true
-	},
-	"extends": ["standard", "plugin:prettier/recommended"],
-	"env": {
-		"mocha": true,
-		"node": true
-	},
-	"plugins": ["havven"],
-	"rules": {
-		"havven/no-assert-revert-without-await": "error",
-		"havven/no-assert-invalid-opcode-without-await": "error",
-		"prefer-arrow-callback": "error",
-		"prefer-const": "error"
-	}
-}
