+++ publish/src/constants.js
@@ -1,0 +1,11 @@
+'use strict';
+
+module.exports = {
+	COMPILED_FOLDER: 'compiled',
+	FLATTENED_FOLDER: 'flattened',
+	CONFIG_FILENAME: 'config.json',
+	SYNTHS_FILENAME: 'synths.json',
+	OWNER_ACTIONS_FILENAME: 'owner-actions.json',
+	DEPLOYMENT_FILENAME: 'deployment.json',
+	ZERO_ADDRESS: '0x' + '0'.repeat(40),
+};
