+++ publish/src/util.js
@@ -1,0 +1,100 @@
+'use strict';
+
+const path = require('path');
+const fs = require('fs');
+const readline = require('readline');
+const { gray } = require('chalk');
+const w3utils = require('web3-utils');
+
+const CONFIG_FILENAME = 'config.json';
+const SYNTHS_FILENAME = 'synths.json';
+const OWNER_ACTIONS_FILENAME = 'owner-actions.json';
+
+const DEPLOYMENT_FILENAME = 'deployment.json';
+
+const toBytes4 = str => w3utils.asciiToHex(str, 4);
+
+const ensureNetwork = network => {
+	if (!/^(kovan|rinkeby|ropsten|mainnet)$/.test(network)) {
+		throw Error(
+			`Invalid network name of "${network}" supplied. Must be one of kovan, rinkeby, ropsten or mainnet`
+		);
+	}
+};
+const ensureDeploymentPath = deploymentPath => {
+	if (!fs.existsSync(deploymentPath)) {
+		throw Error(
+			`Invalid deployment path. Please provide a folder with a compatible ${CONFIG_FILENAME}`
+		);
+	}
+};
+
+// Load up all contracts in the flagged source, get their deployed addresses (if any) and compiled sources
+const loadAndCheckRequiredSources = ({ deploymentPath, network }) => {
+	console.log(gray(`Loading the list of synths for ${network.toUpperCase()}...`));
+	const synthsFile = path.join(deploymentPath, SYNTHS_FILENAME);
+	const synths = JSON.parse(fs.readFileSync(synthsFile));
+	console.log(gray(`Loading the list of contracts to deploy on ${network.toUpperCase()}...`));
+	const configFile = path.join(deploymentPath, CONFIG_FILENAME);
+	const config = JSON.parse(fs.readFileSync(configFile));
+
+	console.log(
+		gray(`Loading the list of contracts already deployed for ${network.toUpperCase()}...`)
+	);
+	const deploymentFile = path.join(deploymentPath, DEPLOYMENT_FILENAME);
+	if (!fs.existsSync(deploymentFile)) {
+		fs.writeFileSync(deploymentFile, JSON.stringify({ targets: {}, sources: {} }, null, 2));
+	}
+	const deployment = JSON.parse(fs.readFileSync(deploymentFile));
+
+	const ownerActionsFile = path.join(deploymentPath, OWNER_ACTIONS_FILENAME);
+	if (!fs.existsSync(ownerActionsFile)) {
+		fs.writeFileSync(ownerActionsFile, JSON.stringify({}, null, 2));
+	}
+	const ownerActions = JSON.parse(fs.readFileSync(ownerActionsFile));
+
+	return {
+		config,
+		configFile,
+		synths,
+		deployment,
+		deploymentFile,
+		ownerActions,
+		ownerActionsFile,
+	};
+};
+
+const loadConnections = ({ network }) => {
+	if (!process.env.INFURA_PROJECT_ID) {
+		throw Error('Missing .env key of INFURA_PROJECT_ID. Please add and retry.');
+	}
+	const providerUrl = `https://${network}.infura.io/v3/${process.env.INFURA_PROJECT_ID}`;
+	const privateKey = process.env.DEPLOY_PRIVATE_KEY;
+	const etherscanUrl =
+		network === 'mainnet'
+			? 'https://api.etherscan.io/api'
+			: `https://api-${network}.etherscan.io/api`;
+
+	const etherscanLinkPrefix = `https://${network !== 'mainnet' ? network + '.' : ''}etherscan.io`;
+	return { providerUrl, privateKey, etherscanUrl, etherscanLinkPrefix };
+};
+
+const confirmAction = prompt =>
+	new Promise((resolve, reject) => {
+		const rl = readline.createInterface({ input: process.stdin, output: process.stdout });
+
+		rl.question(prompt, answer => {
+			if (/y|Y/.test(answer)) resolve();
+			else reject(Error('Not confirmed'));
+			rl.close();
+		});
+	});
+
+module.exports = {
+	toBytes4,
+	ensureNetwork,
+	ensureDeploymentPath,
+	loadAndCheckRequiredSources,
+	loadConnections,
+	confirmAction,
+};
